Make generic Nova testing function.
