+++ circuits/circuit_input_macro/circuit_input_macro_derive/src/lib.rs
@@ -2,7 +2,7 @@ use proc_macro::TokenStream;
 use quote::quote;
 use syn;
 
-#[proc_macro_derive(NovaCircuitInput)]
+#[proc_macro_derive(NovaRoundInput)]
 pub fn circuit_input_macro_derive(input: TokenStream) -> TokenStream {
     let ast = syn::parse(input).unwrap();
     impl_circuit_input_macro(&ast)
@@ -31,7 +31,7 @@ fn impl_circuit_input_macro(ast: &syn::DeriveInput) -> TokenStream {
     });
 
     let gen = quote! {
-        impl NovaCircuitInput for #name {
+        impl NovaRoundInput for #name {
             fn circuit_input(&self) -> std::collections::HashMap<String, serde_json::Value> {
                 let mut map = std::collections::HashMap::new();
                 #(#insertions)*
