+++ circuits/circuit_input_macro/src/lib.rs
@@ -1,3 +1,3 @@
-pub trait NovaCircuitInput {
+pub trait NovaRoundInput {
     fn circuit_input(&self) -> std::collections::HashMap<String, serde_json::Value>;
 }
