+++ circuits/src/lib.rs
@@ -1,3 +1,9 @@
+mod nova;
+
+// State transition
 mod mint;
 mod send;
 mod withdraw;
+
+// Post shutdown distribution
+// mod filter;
