+++ circuits/src/mint.rs
@@ -1,13 +1,12 @@
-use circuit_input_macro::NovaCircuitInput;
-use circuit_input_macro_derive::NovaCircuitInput;
+use crate::nova::{nova, NovaInput};
+use circuit_input_macro::NovaRoundInput;
+use circuit_input_macro_derive::NovaRoundInput;
 use ff::PrimeField;
-use nova_scotia::{
-    circom::reader::load_r1cs, create_public_params, create_recursive_circuit, FileLocation, F1,
-    G2, S1, S2,
-};
-use nova_snark::{traits::Group, CompressedSNARK};
+use nova_scotia::{circom::reader::load_r1cs, FileLocation, F1};
+use pasta_curves::Fq;
 use serde::{Deserialize, Serialize};
-use std::{env::current_dir, time::Instant};
+use serde_json::Value;
+use std::{collections::HashMap, env::current_dir};
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 #[allow(non_snake_case)]
@@ -16,7 +15,23 @@ struct Mint {
     private_inputs: Vec<MintRound>,
 }
 
-#[derive(Serialize, Deserialize, Debug, Clone, NovaCircuitInput)]
+impl NovaInput for Mint {
+    fn initial_inputs(&self) -> Vec<Fq> {
+        vec![
+            F1::from_str_vartime(&self.step_in[0]).unwrap(),
+            F1::from_str_vartime(&self.step_in[1]).unwrap(),
+        ]
+    }
+
+    fn round_inputs(&self) -> Vec<HashMap<String, Value>> {
+        self.private_inputs
+            .iter()
+            .map(|v| v.circuit_input())
+            .collect()
+    }
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone, NovaRoundInput)]
 #[allow(non_snake_case)]
 struct MintRound {
     sender: String,
@@ -28,7 +43,7 @@ struct MintRound {
     pathIndices: Vec<String>,
 }
 
-pub fn nova(iteration_count: usize) {
+pub fn mint(iteration_count: usize) {
     let root = current_dir().unwrap();
 
     let circuit_file = root.join("circuits/build/compiled_circuit/mint.r1cs");
@@ -36,103 +51,7 @@ pub fn nova(iteration_count: usize) {
     let witness_generator_wasm = root.join("circuits/build/compiled_circuit/mint_js/mint.wasm");
 
     let mint_data: Mint = serde_json::from_str(include_str!("../inputs/mint.json")).unwrap();
-
-    println!("json: {mint_data:?}");
-
-    let start_public_input = vec![
-        F1::from_str_vartime(&mint_data.step_in[0]).unwrap(),
-        F1::from_str_vartime(&mint_data.step_in[1]).unwrap(),
-    ];
-
-    let private_inputs = mint_data
-        .private_inputs
-        .iter()
-        .map(|v| v.circuit_input())
-        .collect();
-
-    let pp = create_public_params(r1cs.clone());
-
-    println!(
-        "Number of constraints per step (primary circuit): {}",
-        pp.num_constraints().0
-    );
-    println!(
-        "Number of constraints per step (secondary circuit): {}",
-        pp.num_constraints().1
-    );
-
-    println!(
-        "Number of variables per step (primary circuit): {}",
-        pp.num_variables().0
-    );
-    println!(
-        "Number of variables per step (secondary circuit): {}",
-        pp.num_variables().1
-    );
-
-    println!("Creating a RecursiveSNARK...");
-    let start = Instant::now();
-    let recursive_snark = create_recursive_circuit(
-        FileLocation::PathBuf(witness_generator_wasm),
-        r1cs,
-        private_inputs,
-        start_public_input.clone(),
-        &pp,
-    )
-    .unwrap();
-    println!("RecursiveSNARK creation took {:?}", start.elapsed());
-
-    // TODO: empty?
-    let z0_secondary = vec![<G2 as Group>::Scalar::zero()];
-
-    // verify the recursive SNARK
-    println!("Verifying a RecursiveSNARK...");
-    println!("z0_primary: {start_public_input:?}");
-    println!("z0_secondary: {z0_secondary:?}");
-    let start = Instant::now();
-    let res = recursive_snark.verify(
-        &pp,
-        iteration_count,
-        start_public_input.clone(),
-        z0_secondary.clone(),
-    );
-    println!(
-        "RecursiveSNARK::verify: {:?}, took {:?}",
-        res,
-        start.elapsed()
-    );
-    assert!(res.is_ok());
-
-    // produce a compressed SNARK
-    println!("Generating a CompressedSNARK using Spartan with IPA-PC...");
-    let start = Instant::now();
-    let (pk, vk) = CompressedSNARK::<_, _, _, _, S1, S2>::setup(&pp).unwrap();
-    let res = CompressedSNARK::<_, _, _, _, S1, S2>::prove(&pp, &pk, &recursive_snark);
-    println!(
-        "CompressedSNARK::prove: {:?}, took {:?}",
-        res.is_ok(),
-        start.elapsed()
-    );
-    assert!(res.is_ok());
-    let compressed_snark = res.unwrap();
-
-    // verify the compressed SNARK
-    println!("Verifying a CompressedSNARK...");
-    println!("z0_primary: {start_public_input:?}");
-    println!("z0_secondary: {z0_secondary:?}");
-    let start = Instant::now();
-    let res = compressed_snark.verify(
-        &vk,
-        iteration_count,
-        start_public_input.clone(),
-        z0_secondary,
-    );
-    println!(
-        "CompressedSNARK::verify: {:?}, took {:?}",
-        res.is_ok(),
-        start.elapsed()
-    );
-    assert!(res.is_ok());
+    nova(iteration_count, r1cs, mint_data, witness_generator_wasm);
 }
 
 #[cfg(test)]
@@ -141,6 +60,6 @@ mod tests {
     // use std::env;
     #[test]
     fn test_nova() {
-        nova(2);
+        mint(2);
     }
 }
